// ecoflow/src/models/enums.ts
export enum zonas {
  zona_1 = "Pueblo Nuevo",
  zona_2 = "Barrio Obrero",
  zona_3 = "La Machirí",
  zona_4 = "La Concordia",
  zona_5 = "El Centro",
}

export enum nombresEstaciones {
  estacion_1 = "Eco-Estación Polideportivo",
  estacion_2 = "Eco-Estación Los Mangos",
  estacion_3 = "Eco-Estación UNET",
  estacion_4 = "Eco-Estación Terminal",
  estacion_5 = "Eco-Estación Plaza Bolívar",
}

export enum coordenadasEstaciones {
  estacion_1 = "7.7941,-72.2031",
  estacion_2 = "7.7778,-72.2242",
  estacion_3 = "7.7925,-72.1968",
  estacion_4 = "7.7602,-72.2348",
  estacion_5 = "7.7674,-72.2321",
}

export enum endpoints {
  estaciones = "/api/estaciones/",
  viajes = "/api/viajes/",
  vehiculos = "/api/vehiculos/",
}

export enum datosUsuario {
  coordenadas= "7.75378,-72.22987", 
  hora = "06:30 AM",
}